Update openshs and main_ui
Now it uses by default the current date
