+++ app/main_ui/__init__.py
@@ -1,0 +1,1 @@
+from .menu_bar import *
